Finalize e2e GraphQL tests (#342)
* Remove __typename from selected application fields when parsing query

* Bring back typename test

* Update expected test result against new API

* Update expected result after GraphQL API change

* Add entry to CHANGELOG.md

* Bring back query tests

* Move method into db utilities

* Have some test data

* Remove test
