+++ aquadoggo/src/test_utils/mod.rs
@@ -12,6 +12,7 @@ pub use config::{TestConfiguration, TEST_CONFIG};
 pub use db::{drop_database, initialize_db, initialize_db_with_url};
 pub use helpers::{build_document, doggo_fields, doggo_schema, schema_from_fields};
 pub use node::{
-    add_document, add_schema, populate_and_materialize, populate_store_config, TestNode,
+    add_document, add_schema, add_schema_and_documents, populate_and_materialize,
+    populate_store_config, TestNode,
 };
 pub use runner::test_runner;
